外部ファイルとして読み込む為にclass化とファイル名変更
